Update input arguments to check_file()
Remove keyword argument 'output_file' since it is not used. Changed
keyword argument 'standard_dictionary' to an argument since it is
mandatory and because the funtion crashed when called with the previous
default value of "None". The calls to this function in ags4_cli.py and
test_check.py did not require any modification.