Speed up checks for Rule 10
Iterating over rows in a dataframe using the iterrows() method is quite
slow, therefore the dataframe with errors is first converted to a
dictionary and then iteration done on it. This results in a significant
increase in speed when checking large files (e.g > 100k rows) with a
high number of Rule 10 infringements (e.g. > 10k).