Make it easier to access standard dictionary files
The standard dictionary to be used when checking a file can now be
chosen simply by specifying it's version. This is primarily a shortcut
for convenience and the more general way of specifying any dictionary
file by providing a filepath still works as before.

The CLI and test suite have also been updated accordingly.