Ignore test_batch.py during testing
The test_batch.py file is an attempt to make it easier for users to add
test files and expected results. This is still a work in progress and at
present simply repeats all the units tests as a batch and unnecessarily
consumes CI resources. It will be ignored for now. To be revisited in
the future.