Error fix for: len(line) != len(headings[group])
Working with AGS every day, this error appears a lot in descriptions and remarks (though it can be any text field in the AGS) when the DESC or REM contains a carriage return (line break).

When this error is raised, the AGS4_to_dataframe() function cannot be called - forcing the user to open up the file in a text editor and physically remove the line breaks in the affected text fields.

I've never had this error raised outside of text fields containing carriage returns and it breaks my workflow.