Error fix for: len(line) != len(headings[group])
During everyday use of AGS, this error appears a lot in descriptions and remarks (though it can be any text field in the AGS) when the DESC or REM contains a carriage return (line break).

When this error is raised, the AGS4_to_dataframe() function cannot be called - forcing the user to open up the file in a text editor and physically remove the line breaks in the affected text fields.

I haven't encountered this error outside of text fields containing carriage returns and its cumbersome to fix, so I've implemented a solution in place of the current function, which currently prints the error and exits.

This will concatenate the line where the error occurred with the line above it, save over the file and re-open it, proceeding in a loop until all linebreaks are removed, printing the AGS4 error, along with the heading & data for each error.