Remove latin1/iso-8859-1 as an encoding option in the CLI
If 'latin1' encoding is inadvertently used to check a file with
'cp1252' encoding, it will not flag characters 128-159 as invalid. On
the other hand, 'cp1252' can be safely used to check files encoded in
both 'cp1252' and 'latin1' as it is a superset of 'latin1'. Removing the
'latin1' option avoids this pitfall by forcing the use of 'cp1252'.