Make convert_to_numeric() more robust
The function returns a ValueError exception when the TYPE row contains a
value that is not a string because the str.contains() method returns NaN
values which cannot be used in a mask. This scenario is rare but it can
happen when convert_to_numeric() is applied on a dataframe with line
numbers (returned by AGS4_to_dataframe() function called with
get_line_numbers=True). This exception can be avoided by setting
na=False when the str.contains() method is called.