Avoid checking significant figures in zeros
Significant figures cannot be determined for zeros, therefore they are
explicitly ignored in rule_8() check for SF. The existing code relied on
AGS4._format_SF() to ignore zeros, but it returns "0.0" when round
tripping from text to numeric to text again, regardless of the value in
the input file (e.g. all of "0", "0.0", "0.00" will get returned as
"0.0"). As a result, any zero value that is not 0.0 is flagged as having
the wrong number of significant figures.