Fix #105: Add 'General' message to error if exception is raised
The previous behavior of the check_file() function was to allow
exceptions to be raised. However, this interfered with the output of the
function and stopped the error log from being created. A 'General'
message is now be added when an exception is caught to make it clear to
the user that the validation terminated prematurely.