ci: remove tags that interferred with gitlab runner
The tag 'linux' stopped the allocation of a runner to execute the CI
pipeline because it did not match with the tags of any available
runners. This was not a problem until now, so I suspect that the latest
GitLab version enforces an exact match of tags. Removing this tag
solved the issue.